services_demoplan-unified_1759943521.85486-5058385db2bb48e0a515d2457909c93a/requirements.txt
# DemoPLAN Phase 1 Requirements - Core Chat + File Upload + OCR
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Google Cloud & Firebase
google-cloud-firestore==2.13.1
google-auth==2.23.4
google-cloud-storage>=2.10.0

# NEW: Document AI for OCR
google-cloud-documentai>=2.20.0

# AI & LLM Services
google-generativeai==0.8.0

# File Processing & DXF
ezdxf==1.1.0

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3

# Data Processing
numpy==1.25.2
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.0

# Async & HTTP
httpx==0.25.2
aiofiles==23.2.1

# Utilities
python-dateutil==2.8.2

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1